add fav icon fix header menu fix in migration laradock

+++ database/migrations/2017_11_26_013050_add_user_role_relationship.php
@@ -14,7 +14,7 @@ class AddUserRoleRelationship extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->unsignedBigInteger('role_id')->change();
+            $table->unsignedInteger('role_id')->change();
             $table->foreign('role_id')->references('id')->on('roles');
         });
     }
@@ -31,7 +31,7 @@ class AddUserRoleRelationship extends Migration
         });
 
         Schema::table('users', function (Blueprint $table) {
-            $table->unsignedBigInteger('role_id')->change();
+            $table->unsignedInteger('role_id')->change();
         });
     }
 }
